Add additional features to parser (#280)
This PR mostly completes the reader and data structures for terms and
typedefs with respect to the OBO flat file format.

It's still not quite accurate in the way it represents instances, but
this is hard to deal with since ROBOT cowardly refuses to work with
instances in OBO format. The functional OWL PR will make it easier to
test this