Improve WikiPathways ingestion (#416)
This PR adds the following:

1. Adds subclass relations from WikiPathways terms to PW terms that
they're annotated on. If none are given, annotates directly on to the
top level PW pathway term
2. Adds RO 'contributes to condition' annotation properties to all
diseases attached to each WP entry. This is a bit presumptions, since
this is a causal relation, but RO doesn't have anything else for
associations at the moment.
3. Adds RO 'located in' annotation properties to cells tagged on each WP
entry
4. Adds image depictions, which OLS should be able to pick up